requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
beautifulsoup4==4.10.0
black==22.1.0
bs4==0.0.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
discord.py==1.7.3
feedparser==6.0.8
idna==3.3
multidict==6.0.2
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
python-dotenv==0.19.2
requests==2.27.1
sgmllib3k==1.0.0
soupsieve==2.3.1
tomli==2.0.1
typing_extensions==4.1.0
urllib3==1.26.8
yarl==1.7.2
